diabetic_retinopathy/requirements.txt
gin-config==0.3.0
tensorflow==2.3.1
tensorflow-datasets==4.1.0
tensorflow-addons==0.11.2
pandas==1.1.4
matplotlib==3.3.3
seaborn==0.11.0